feat(lobby): move / to /join
